\name{Autograd-operator}
\alias{Autograd-operator}
\alias{+.autograd}
\alias{-.autograd}
\alias{*.autograd}
\alias{/.autograd}
\title{Operator overloading by autograd class}
\usage{
e1 + e2

e1 - e2

e1 * e2

etc.
}
\description{
Many operators are overloaded by the autograd class. Using smart bookkeeping and differential algebra rules, higher-order derivatives are also supported.
}
\seealso{
\link{deriv}
}
