\docType{package}
\name{Autograd-package}
\alias{Autograd}
\alias{Autograd-package}
\title{Automatic differentation in R}
\description{
Native implementation of automatic differentation. Implemented as S3 methods on class \code{Autograd}. Works with derivatives of any order.
}
\details{
Mix scalars, vectors, array etc. as usual in R. Does NOT contain any optimizer, and is not made for speed.
}
\seealso{
\link{Autograd-class}
}